Make PublishBuilder::packet_id() take an Option
Since the packet ID is optional in a PUBLISH packet depending on the
QoS, adapter functions that construct a PUBLISH packet may need to
conditionally insert a packet ID. With the current builder method, this
is not ergonomic -- the user must create a mutable binding for the
builder and conditionally call the `packet_id()` method. By changing the
method to take an Option, the user can unconditionally call the method
and avoid interrupting the builder pattern. See the simplification in
`examples/publish.rs` as an example.